pyproject.toml
[project]
name = "macsetup"
version = "1.0.0"
description = "macOS Configuration Sync CLI - Capture, store, and restore macOS machine configurations"
readme = "README.md"
requires-python = ">=3.14"
license = "MIT"
authors = [{ name = "atyagi" }]
keywords = ["macos", "dotfiles", "homebrew", "configuration", "sync"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python :: 3.14",
    "Topic :: System :: Installation/Setup",
    "Topic :: Utilities",
]

dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.20",
]

[project.scripts]
macsetup = "macsetup.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/macsetup"]

[tool.ruff]
target-version = "py314"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "SIM117", # nested with statements (intentional in tests for readability)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.14",
]
